Use integers instead of floats for custom leaderboards

+++ DevilDaggersWebsite/Code/Database/CustomLeaderboards/CustomEntry.cs
@@ -16,7 +16,7 @@ namespace DevilDaggersWebsite.Code.Database.CustomLeaderboards
 		[ForeignKey("CustomLeaderboardId")]
 		public CustomLeaderboard CustomLeaderboard { get; set; }
 
-		public CustomEntry(int playerId, string username, float time, int gems, int kills, int deathType, int shotsHit, int shotsFired, int enemiesAlive, int homing, float levelUpTime2, float levelUpTime3, float levelUpTime4, DateTime submitDate, string clientVersion)
+		public CustomEntry(int playerId, string username, int time, int gems, int kills, int deathType, int shotsHit, int shotsFired, int enemiesAlive, int homing, int levelUpTime2, int levelUpTime3, int levelUpTime4, DateTime submitDate, string clientVersion)
 			: base(playerId, username, time, gems, kills, deathType, shotsHit, shotsFired, enemiesAlive, homing, levelUpTime2, levelUpTime3, levelUpTime4, submitDate, clientVersion)
 		{
 		}
